move get_reference_surface-temperatures to prescriber
